requirements.txt
Sanic==19.3.1
Pillow==5.4.1
BeautifulSoup4==4.7.1
requests==2.21.0
SQLAlchemy==1.3.2
asyncpg==0.18.3
asyncio==3.4.3
psycopg2==2.8.1
